+++ examples/scan.rs
@@ -1,4 +1,4 @@
-use conifer::*;
+use conifer::prelude::*;
 
 fn main() {
     let c = Config::auto();
