+++ src/prelude.rs
@@ -1,4 +1,4 @@
 pub use crate::gesture::Gesture;
-pub use crate::Config;
-pub use crate::RunResponse;
+pub use crate::config::Config;
+pub use crate::config::RunResponse;
 pub use std::error::Error;
